Added 'algorithm.py' file
Added a file called 'algorithm.py'. Right now, it's just a text outline of what I plan to code for the algorithm. I still need to decide + implement a game that I want to use RL for. Alternatively, I might try to make it a generalized algorithm for any game, but that might be difficult because it isn't easy to picture a universalizable way to store the moves and states in a matrix (which is why I might use a hashtable instead, since that can be indexed with state+move, if I have a string representation of each)